color: _Float16 as the channel lingua franca; RGBA16F tiles + target
cnvs_rgba channels become _Float16, colour tiles become RGBA16F, and the
compositor target becomes MTLPixelFormatRGBA16Float; blending happens in
half/float.  8-bit appears only at the edges the Canvas spec mandates it:
getImageData, putImageData, and PNG (the compositor quantizes the 16F target on
readback; putImageData stays an 8-bit replace).  Same colour space as before --
this is the drop-in step, not linear-light; a full-image diff of the gallery
shows a max per-channel delta of 1 (pure rounding).

Float -> _Float16 narrowing is funnelled through one helper, cnvs_rgba_of, so
-Wimplicit-float-conversion has exactly one (explicit) place to live; the
gradient lerp computes in float and narrows once. _Float16 was otherwise
friction-free under -Weverything (the only gotcha is the same -Wbad-function-cast
rule we already follow: never cast a call result directly).

The payoff at 8-bit output is small by design -- the win is the working space:
no 8-bit rounding accumulating across stacked composites, headroom past [0,1],
and the foundation for linear-light compositing and wide-gamut/HDR later.

+++ src/compositor.h
@@ -6,8 +6,9 @@
 // A Metal backend implements this today; nothing here is GPU-specific, so a pure
 // CPU backend could implement the same ABI.
 //
-// Tiles are tightly packed straight-alpha RGBA8, row-major, top row first.  All
-// regions must lie within the target (the caller clips to it).
+// Colour tiles are tightly packed straight-alpha RGBA16F (_Float16 channels),
+// row-major, top row first; putImageData tiles are RGBA8.  All regions must lie
+// within the target (the caller clips to it).
 
 #include <ptrcheck.h>
 #include <stdint.h>
@@ -24,13 +25,14 @@ void compositor_destroy(compositor *__single c);
 void compositor_set_clip(compositor *__single c,
                          uint8_t const *__counted_by(len) mask, int len);
 
-// Source-over a w*h tile at (x,y), multiplied by the current clip mask.  This is
-// every painted fill and stroke (colour and coverage already baked into alpha).
+// Source-over a w*h RGBA16F tile at (x,y), multiplied by the current clip mask.
+// This is every painted fill and stroke (colour and coverage already baked into
+// alpha).
 void compositor_blend(compositor *__single c, int x, int y, int w, int h,
-                      uint8_t const *__counted_by(w * h * 4) tile);
+                      _Float16 const *__counted_by(w * h * 4) tile);
 
-// Overwrite a w*h region at (x,y) with the tile (no blend, ignores the clip).
-// This is putImageData.
+// Overwrite a w*h region at (x,y) with the RGBA8 tile (no blend, ignores the
+// clip).  This is putImageData, whose source is 8-bit.
 void compositor_replace(compositor *__single c, int x, int y, int w, int h,
                         uint8_t const *__counted_by(w * h * 4) tile);
